+++ packages/thumbprint-react/utils/use-close-on-escape.js
@@ -10,21 +10,18 @@ const EVENT_NAME = 'keyup';
  * @param {Boolean} isActive Whether or not the key listener does anything.
  */
 export default function useCloseOnEscape(doClose, isActive = false) {
-    useEffect(
-        () => {
-            const handleKeyUp = event => {
-                if (isActive && event.keyCode === ESC_KEY) {
-                    event.preventDefault();
-                    doClose();
-                }
-            };
+    useEffect(() => {
+        const handleKeyUp = event => {
+            if (isActive && event.keyCode === ESC_KEY) {
+                event.preventDefault();
+                doClose();
+            }
+        };
 
-            document.addEventListener(EVENT_NAME, handleKeyUp);
+        document.addEventListener(EVENT_NAME, handleKeyUp);
 
-            return () => {
-                document.removeEventListener(EVENT_NAME, handleKeyUp);
-            };
-        },
-        [doClose, isActive],
-    );
+        return () => {
+            document.removeEventListener(EVENT_NAME, handleKeyUp);
+        };
+    }, [doClose, isActive]);
 }