+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+-   [Patch] Code formatting tweaks due to Prettier version bump.
+
+## 9.1.1 - 2019-08-23
+
+### Fixed
+
+-   [Patch] Prevent noscript image when `forceEarlyRender` prop is used.
+
 ## 9.1.0 - 2019-08-22
 
 ### Added